requirements.txt
gradio>=4.36.1
python-dotenv>=1.0.1
openai>=1.40.0
sentence-transformers>=2.7.0
scikit-learn>=1.5.1
yake>=0.4.8
networkx>=3.3
numpy>=1.26.4
matplotlib>=3.8.0
